+++ lib/SignupPage.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:monew/ProfilePage.dart';
+import 'package:monew/utils/utils.dart';
 
 class SignupPage extends StatefulWidget {
   const SignupPage({Key? key}) : super(key: key);
@@ -22,52 +23,6 @@ class _SignupState extends State<SignupPage> {
         logo(),
     ],)
     ),
-    );
-  }
-}
-
-class title extends StatelessWidget{
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        SizedBox(height: 70),
-        Text(
-          'MoNew',
-          textAlign: TextAlign.center,
-          style: TextStyle(
-            color: Colors.black,
-            fontSize: 24,
-            fontFamily: 'Inter',
-            fontWeight: FontWeight.w600,
-            height: 0.06,
-            letterSpacing: -0.24,
-          ),
-        ),
-        SizedBox(height: 100),
-        Container(
-          width: 80,
-          height: 27,
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            mainAxisAlignment: MainAxisAlignment.start,
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              Text(
-                '회원가입',
-                textAlign: TextAlign.center,
-                style: TextStyle(
-                  color: Colors.black,
-                  fontSize: 18,
-                  fontFamily: 'Inter',
-                  fontWeight: FontWeight.w600,
-                  height: 0.08,
-                ),
-              ),
-            ],
-          ),
-        ),
-      ],
     );
   }
 }
@@ -132,10 +87,17 @@ class _InputState extends State<SignupInput>{
                               fontSize: 14,
                               fontFamily: 'Inter',
                               fontWeight: FontWeight.w400,
-                              height: 0.10,
+                              height: 1.2,
                             ),
                             border: InputBorder.none,
-                          )
+                          ),
+                        style: TextStyle(
+                          color: Colors.black,
+                          fontSize: 14,
+                          fontFamily: 'Inter',
+                          fontWeight: FontWeight.w400,
+                          height: 1.2,
+                        ),
                       ),
                     ),
                   ),
@@ -171,7 +133,7 @@ class _InputState extends State<SignupInput>{
                         fontSize: 14,
                         fontFamily: 'Inter',
                         fontWeight: FontWeight.w500,
-                        height: 0.10,
+                        height: 1.2,
                       ),
                       shape: RoundedRectangleBorder(
                         borderRadius: BorderRadius.circular(8),
@@ -184,7 +146,7 @@ class _InputState extends State<SignupInput>{
                         fontSize: 14,
                         fontFamily: 'Inter',
                         fontWeight: FontWeight.w500,
-                        height: 0.10,
+                        height: 1.2,
                       ),
                     ),
                   ),
@@ -234,10 +196,17 @@ class _InputState extends State<SignupInput>{
                               fontSize: 14,
                               fontFamily: 'Inter',
                               fontWeight: FontWeight.w400,
-                              height: 0.10,
+                              height: 1.2,
                             ),
                             border: InputBorder.none,
-                          )
+                          ),
+                        style: TextStyle(
+                          color: Colors.black,
+                          fontSize: 14,
+                          fontFamily: 'Inter',
+                          fontWeight: FontWeight.w400,
+                          height: 1.2,
+                        ),
                       ),
                     ),
                   ),
@@ -287,10 +256,17 @@ class _InputState extends State<SignupInput>{
                               fontSize: 14,
                               fontFamily: 'Inter',
                               fontWeight: FontWeight.w400,
-                              height: 0.10,
+                              height: 1.2,
                             ),
                             border: InputBorder.none,
-                          )
+                          ),
+                        style: TextStyle(
+                          color: Colors.black,
+                          fontSize: 14,
+                          fontFamily: 'Inter',
+                          fontWeight: FontWeight.w400,
+                          height: 1.2,
+                        ),
                       ),
                     ),
                   ),
@@ -340,10 +316,17 @@ class _InputState extends State<SignupInput>{
                               fontSize: 14,
                               fontFamily: 'Inter',
                               fontWeight: FontWeight.w400,
-                              height: 0.10,
+                              height: 1.2,
                             ),
                             border: InputBorder.none,
-                          )
+                          ),
+                        style: TextStyle(
+                          color: Colors.black,
+                          fontSize: 14,
+                          fontFamily: 'Inter',
+                          fontWeight: FontWeight.w400,
+                          height: 1.2,
+                        ),
                       ),
                     ),
                   ),
@@ -376,7 +359,7 @@ class _InputState extends State<SignupInput>{
                     fontSize: 14,
                     fontFamily: 'Inter',
                     fontWeight: FontWeight.w500,
-                    height: 0.10,
+                    height: 1.2,
                   ),
                   shape: RoundedRectangleBorder(
                     borderRadius: BorderRadius.circular(8),
@@ -389,7 +372,7 @@ class _InputState extends State<SignupInput>{
                     fontSize: 14,
                     fontFamily: 'Inter',
                     fontWeight: FontWeight.w500,
-                    height: 0.10,
+                    height: 1.2,
                   ),
                 ),
               ),
@@ -418,7 +401,7 @@ class _InputState extends State<SignupInput>{
                       fontSize: 14,
                       fontFamily: 'Inter',
                       fontWeight: FontWeight.w400,
-                      height: 0.10,
+                      height: 1.2,
                     ),
                   ),
                   const SizedBox(width: 8),
@@ -483,7 +466,7 @@ class _InputState extends State<SignupInput>{
                             fontSize: 14,
                             fontFamily: 'Inter',
                             fontWeight: FontWeight.w500,
-                            height: 0.10,
+                            height: 1.2,
                           ),
                         ),
                       ],
@@ -497,27 +480,6 @@ class _InputState extends State<SignupInput>{
           ],
         ),
       ),
-    );
-  }
-}
-
-class logo extends StatelessWidget {
-  @override
-  Widget build(BuildContext context){
-    return Column(
-      children: [
-        SizedBox(height: 50),
-        Container(
-          width: 70,
-          height: 67,
-          decoration: BoxDecoration(
-            image: DecorationImage(
-              image: AssetImage('assets/images/monew_logo_bottom.png'),
-              fit: BoxFit.fill,
-            ),
-          ),
-        ),
-      ],
     );
   }
 }
